Fix splitter auto-centering logic
It seems wrong to auto-center the splitter when m_dontCenterSplitter is
set, but this is what happened before, as long as there were no children
properties.

Don't do any more, but do call SetSplitterLeft() as long as the splitter
hadn't been explicitly positioned, as this seems the most useful
behaviour.