Remove wxLIST_COMPATIBILITY define which was always on
This was somewhat confusing as "compatibility" here was meant to be in
the sense of "compatibility with the impossibly ancient, wx 1.x wxList"
and not "compatibility with the reasonably ancient wx 2.8 wxList", so
just remove this symbol instead of trying to come up with a better name
for it.

The functions guarded by it are all already deprecated and give warnings
when used and may be removed at some later date as all the other
deprecated functions anyhow.

No real changes.