Change theme when changing appearance in wxGTK if necessary
At least under Ubuntu 22.04 (GNOME 42) resetting the dark preference is
not enough to switch to light appearance when the theme is set to the
default value of "Yaru-dark" in system-wide dark mode, so change the
theme to remove the explicit dark theme selector to make this work.

Note that it does not seem to be necessary to add "-dark" back when
switching from light mode to dark, i.e. using "prefer-dark-theme" to
choose the dark mode of the default theme seems to have exactly the same
effect as choosing the dark theme variant.

Closes #24476.