CYGWIN: DLL files must be installed into bin.
I compiled successfully wxWidgets for CYGWIN with CMake.
However, it happened that DLL files are installed into the lib directory instead of bin.
Hopefully, the fix is very easy because runtime_dir just needs to be set to bin when CYGWIN is detected, exactly like WIN32.
This PR fixes this issue.