Don't draw wxStaticText as disabled in dark mode if it isn't
The change of bc72a5d003 (Fix drawing of disabled wxStaticText in dark
mode, 2023-05-18) resulted in drawing wxStaticText using disabled colour
even when wxStaticText itself wasn't actually disabled, e.g. it appeared
in this colour while a modal dialog using the window containing it as
the owner was shown because this disabled the parent window -- but it
shouldn't change the wxStaticText colour in dark mode, just as it
doesn't grey it out in the default light one.

Simply use IsThisEnabled() instead of IsEnabled() to only use the custom
drawing hack for the disabled controls in this case.