Always restore wxStaticText original m_foregroundColour
It's not clear if changing it to the "disabled" colour in dark mode
if m_hasFgCol was false actually resulted in any problems, but it seems
safer to always restore it after drawing the control using the disabled
colour in any case.

This corrects another (possible) problem after the changes of bc72a5d003
(Fix drawing of disabled wxStaticText in dark mode, 2023-05-18).