Move tab border pen setting outside of DrawNotebookTab()
No real changes, just set the pen only once instead of doing it in each
DrawNotebookTab(): this is a micro-optimization but, more importantly,
will allow draw the tab separator line outside of DrawNotebookTab()
using the same pen in upcoming commits.

Also don't draw anything at all if there are no tabs to avoid setting
the pen uselessly in this case.