Fix wxTE_PROCESS_ENTER in wxGTK comboboxes with autocomplete
This style was reset and never re-enabled because the comboboxes get
multiple grab-notify notifications, unlike text controls, which only get
one.

Fix this by adding a counter of the number of times grab-notify was
called indicating that the popup was shown or hidden and only update the
style when its status actually effectively changes.

Also use a more clear name for the function called when the popup is
shown as having a Toggle(bool toggle) function was a bit confusing.

Closes #24489.

Closes #24394.