Don't set focus to hidden wxSimplebook current page
Unconditionally calling SetFocus() on the current page in wxSimplebook
resulted in a weird situation with focus in wxGTK as the hidden window
had focus at wx level but not at GTK one, where focus was just lost.

For now, work around this by avoiding calling SetFocus() in this case,
even though it might be better to refuse to set focus to a hidden window
in SetFocus() itself.