Guard against re-entering in wxOSX wxDropSource::DoDragDrop()
At least when using CEF, repeatedly starting drag-and-drop operation may
result in stack overflow.

Messages processed by CEF are handled when wxEVT_IDLE is triggered and
the problematic scenario is when:

1. Cocoa calls OSXDefaultModeObserverCallBack()
2. wxEVT_IDLE is triggered.
3. wxWebViewChromium::OnIdle() is called.
4. CefDoMessageLoopWork() performs various jobs for CEF.

If DoDragDrop() is invoked during CefDoMessageLoopWork(), and there is a
recursive call to it, then the previous DoDragDrop() can't return
anymore, and will, in turn, prevent CefDoMessageLoopWork() from
returning to the caller.

Prevent recursively re-entering this function to avoid this.

See #24520.

Co-authored-by: Vadim Zeitlin <vadim@wxwidgets.org>