Use actual timestamps in wxOSX messages to avoid crash with CEF
Change the two dummy messages to use systemUpTime as timestamp instead
of 0 to fix CEF renderer process crash.

Beginning with Chromium 97, EventLatencyTracingRecorder was added to
trace the performance. And the timestamp 0 used will lead to a check
failure in EventLatencyTracingRecorder::RecordEventLatencyTraceEvent
see:

https://github.com/chromium/chromium/blob/109.0.5414.120/cc/metrics/event_latency_tracing_recorder.cc

Way to reproduce the crash:
Right click to popup the context-menu created by CEF.
Then the CEF renderer process will crash.

Set the timestamp to [[NSProcessInfo processInfo] systemUptime] to fix it.

See #24520.