Fix creating controls on startup in the widgets sample under Mac
This got broken by a8b4753d1d (Reconnect events to the widget after
recreating it in the sample, 2024-02-18) because checking whether the
children list is empty doesn't work correctly in wxOSX (and wxUniv)
because scrollbars may be present in this list even if there are no
other children yet.

Fix this by checking if we have any "real" children explicitly.

Closes #24528.

Closes #24529.