Fix wxGridBagSizer layout if horizontal and vertical gaps differ
They were mixed up when computing the item size, yielding wrong results
when their values were different.

Fix this and also use m_[hv]gap directly for consistency with the rest
of the code in this file instead of using Get[HV]Gap() accessors that
were not used anywhere else.

Closes #24550.