Do not scale the size of bitmap in wxMSW Create(size, dc) overload
The size passed to Create() should be interpreted in the same way as the
size passed to CreateWithLogicalSize() and in wxMSW this means that it
should _not_ be scaled.

This makes the size of the bitmap returned by this function overload
consistent with wxDC::GetSize(), so it looks like the right thing to do,
even if it's a backwards-incompatible change.