BUG FIX #24531 & 24535
Initial commit modified:
 - constexpr instead of const
 - removal of iostream
 - wxMax instead of MAX
 - HasFlag() instead of style & ...

I have still kept all modifications in one commit because the old source code is mostly a copy & paste of old Carbon code that does not make much sense for Cocoa (it worked but that's it). All these modifications are necessary to bring the code from Carbon to Cocoa.
Some further modifications can be done to clean up the code further. But this commit lets actually the sliders look like sliders on Cocoa.