wxrc: Generate C++ code which is faster to compile and smaller
Avoiding performing the conversions from "const wxChar*" to wxString in
all places where XRC_ADD_FILE is used and doing it in just a single
function allows to significantly reduce both the compilation time and
the size of the generated machine code.

Also add a simple script allowing to generate the XRC file that can be
used to test this improvement.

Closes #24579.