[wxGTK][localization] Fix untranslated btn labels
The UI items that have a standard ID and labels identical to to the stock ones
are assigned with GTK stock labels, see e.g. wxButton::SetLabel() in wxGTK.
The stock labels do not get translated if msgctxt field is present in the
translation catalog.
This is reproducible with multiple (all?) languages.

As stock labels are deprecated since GTK 3.10, fix this by reverting
official wx commit 29bd2c9 where stock usage is prolonged until GTK 4.

Issue #24366