wxQt: Fix generating wxEVT_TREE_SEL_{CHANGING,CHANGED} events
The wxEVT_TREE_SEL_CHANGING event is now generated before currentItemChanged()
signal is emitted so we will have the opportunity to not emit the signal altogether
if the wx event is vetoed.

The wxEVT_TREE_SEL_CHANGED event is now generated when the last itemSelectionChanged()
signal is emitted, in which case wxTreeCtrl::GetSelections() is guaranteed to return
the correct selected items if called from the event handler.

Also fix a bug when wxTreeCtrl::EndEditLabel() is called in a control with wxTR_MULTIPLE
flag resulting in GetSelections() returning the old selected item instead of the new one.

And remove a redundant call to setEditTriggers().