build: fix wx-config symlink in out-of-tree installs
When we install out-of-tree, i.e. with a non-empty DESTDIR, the
wx-config symlink is broken: the target of the symlink is not rooted
in DESTDIR.

It would seem the obvious solution would be to prefix the target of
the symlink with $(DESTDIR) and be done with it. That would solves
cross-compilation cases, but would not fix native build when the
install is done in a staging area to prepare a binary package (e.g.
Linux distributions à-la Debian do that for example).

A way to make the symlink work in both cases is by making it relative.
ln has been supporting a --relative option since coreutils 8.16,
released 2012-03-26, more than 12 years ago now; it is relatively safe
to assume that distributions of today do have an ln that supports that
flag.

For systems that do not support that flag (e.g. *BSD, MacOS), fallback
to creating a non-relative symlink, on the assumpion [0] that on those
systems, DESTDIR is most probably never used. Eventually, in case
creating both symlinks fails, keep doing a copy as the ultimate
fallback.

[0] https://github.com/wxWidgets/wxWidgets/pull/24622#issuecomment-2184047170

Note: Makefile.in was manually edited to mirror the bakefile; however,
there was a discrepancy between the two: the bakefile used an uppercase
$(LIBDIR), while the Makefile.in used a lowercase $(libdir). This change
keeps the discrepancy intact.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>