Fix crash on static page of widgets sample
Attempting to change the style or alignment of the static text in the
widgets sample made the sample crash. The bug cause was the same as in
7c125d3 (Fix crash on spin page of widgets sample, 2024-04-18), i.e.,
a8b4753 (Reconnect events to the widget after recreating it in the sample,
2024-02-19).

Fix the issue by calling NotifyWidgetRecreation() for each control only
after they were all recreated.

Closes #24656.

Closes #24660.