Declare a class in WX_DECLARE_HASH_MAP() too
Instead of just using a type alias, define a class inheriting from
std::unordered_map<> in this macro, as this makes it compatible with the
existing code not using wxUSE_STD_CONTAINERS which forward-declares hash
map classes.

And as this was already done in WX_DECLARE_HASH_SET(), it's also more
consistent.