wxQt: Stop overriding wxWindow::GetHandle() in derived classes
There is no need to do so because assigning to m_qtWindow is enough to return a
valid handle.

There is also no need to store an additional pointer just to preserve the actual
type of the handle which can be obtained using a simple cast.

There are no real changes other than simplifying and normalizing the code base.