Fix adding smaller bitmap versions to wxBitmapBundle
Smaller bitmap was inserted in a wrong place, resulting in the entries
array not being sorted by bitmap size any longer and completely breaking
the logic of GetBitmap() which relied on it being sorted.

In particular, this meant that we kept recreating smaller version of the
first bitmap in the bundle again and again when the function was called
for the bitmap of the same size, which resulted in unbounded resource
consumption and, eventually, a crash.

Fix this by inserting the bitmap in the right place.

Note that this commit doesn't include the tests added on master as they
use a new function which doesn't exist in this branch.

See #24703.

(cherry picked from commit 6286fc5ca040697a4fa6df5dc9c56006a57c285c)