Fix adding smaller bitmap versions to wxBitmapBundle
Smaller bitmap was inserted in a wrong place, resulting in the entries
array not being sorted by bitmap size any longer and completely breaking
the logic of GetBitmap() which relied on it being sorted.

In particular, this meant that we kept recreating smaller version of the
first bitmap in the bundle again and again when the function was called
for the bitmap of the same size, which resulted in unbounded resource
consumption and, eventually, a crash.

Fix this by inserting the bitmap in the right place and add a unit test
checking that the number of GDI objects remains more or less constant
even after creating a lot of bitmaps.

Closes #24703.