Fix crash when copying wxBitmaps in wxMSW if DIB creation fails
This shouldn't normally happen, but if it does, e.g. because all the GDI
resources have been exhaused, we still shouldn't crash, so check for
this.

Also add a (disabled by default as it could be potentially dangerous to
run it) test checking that we don't crash in this case.

See #24703.