Really fix wxBitmapBundleImpl::GetIndexToUpscale() return value
The changes of 538eafc78b (Fix bug with wrong GetIndexToUpscale() return
value, 2022-06-03) were wrong but happened to work when there were no
down-scaled versions of the bitmap in the bundle.

But in general, we shouldn't use the previous index, but the index just
before the one with the most appropriate scale, so do this, and extend
the unit test to check for this scenario too -- previously it would fail
due to an assert failure when we tried to upscale a previously
downscaled bitmap.

(cherry picked from commit 31f59355a0b2938cae2137e30149e6391b6fab60)