Fix and simplify check for "real" children in wxWindow
The old code was wrong because it only checked for the scrollbars, which
appear in the list of children under Mac, but not for the "grow box",
which is also there and should be similarly ignored when computing the
best window size.

It was also inefficient because it used dynamic cast and iterated over
the list of children twice, first in wxHasRealChildren() and then in
DoGetBestSize() itself.

Fix this by removing wxHasRealChildren() entirely and using
IsClientAreaChild() in DoGetBestSize() to check for the children to
skip.

Replaces #24566.

See #24708.

Closes #24565.