Add private header wrapping GDK headers inclusion
This allows to have the code checking for the various backends in one
place only and also ensures that we can disable -Wundef in gdkwayland.h
just once instead of doing it in every place where it is included,
because the changes of aba1c66575 (Suppress harmless -Wundef in Wayland
headers, 2024-07-16) were insufficient to fix the FreeBSD CI build, as
the same warning still happens elsewhere.

See #24709.