Add wxInfoDC allowing to query device context metrics
This class is meant to transparently replace the use of wxClientDC in
the existing code where the latter is only used for getting information
about the device context and not for drawing on it, as only such use of
wxClientDC is still guaranteed to work on all supported platforms.

In order to make it impossible to call functions drawing on the DC on an
object of wxInfoDC class, wxInfoDC does not inherit from wxDC. However,
to still allow having functions taking any kind of DC, including
wxInfoDC, provided they don't draw on it, a new wxReadOnlyDC is also
added as a base class of wxDC, meaning that wxDC objects are convertible
to wxReadOnlyDC, and wxInfoDC is also implicitly convertible to it via
conversion operators.

The goal of this design, other than to provide wxInfoDC at all, is to
make sure that applications can use wxReadOnlyDC and wxInfoDC while
still remaining compatible with the previous wxWidgets versions if they
also include fallback definitions for them:

	#ifndef wxHAS_INFO_DC
	    using wxReadOnlyDC = wxDC;
	    using wxInfoDC = wxClientDC;
	#endif