Add wxWindow::PrepareReadOnlyDC() similar to PrepareDC()
Unfortunately PrepareDC() itself can't be changed to take wxReadOnlyDC
instead of wxDC, as this would break existing code overriding this
function, so add a new function taking wxReadOnlyDC, which allows to
call it with wxInfoDC, instead.

Override it in wxScrollWindow to forward to DoPrepareReadOnlyDC(), which
is now implemented in wxScrollHelperBase in addition to and instead of
DoPrepareDC().