Replace wxClientDC with wxInfoDC when possible
Stop using wxClientDC just for measuring string extents and use wxInfoDC
instead. This requires changing many functions to accept wxReadOnlyDC
instead of wxDC, which is backwards-incompatible change as it breaks the
code overriding these functions.

The remaining occurrences of wxClientDC are actually used for drawing on
screen and will need to be dealt with later.