Do not malloc zero bytes in wxInputStream::AllocSpaceWBack
Running the tests, specifically TestExecute, under ElectricFence reveals
that a malloc(0) call is made in wxInputStream::AllocSpaceWBack.

While legal, allocating zero bytes is unusual. Any memory pointed must
not be read or written.

It is likely better to return nullptr in cases when a buffer of
zero bytes was requested.