Fix incorrect handling of ligatures in wxStyledTextCtrl in wxGTK
The clusters in Pango's LayoutIter treat ligatures as one unit.
Thus, if two characters were displayed as a ligature, we would not
be able to move the caret in between the two characters.

Therefore, we calculate the partial text extents by iterating over
every character (except the first one), check if a cursor could
appear in front of it (i.e. this is a grapheme boundary), and if
it could, we add the position to the array. If there has been at
least one character, we also add the position behind the last one
to the array.

Closes #13168.

Closes #24748.