Use darker default background in dark mode in wxMSW
wxSYS_COLOUR_BTNFACE is not appropriate in dark mode, as it's not dark
enough and also inconsistent with the other windows already using
wxSYS_COLOUR_WINDOW for their background, so use the latter colour for
everything by default.

Note that it might make sense to always use wxSYS_COLOUR_WINDOW by
default, as it's supposed to be the default background window colour,
but this could well break something after many years of using a
different default, so avoid changing this, at least for now.