Disallow overriding wxBitmap::Create() outside of wxWidgets
This doesn't really make much sense, this function is only virtual in
the first place to force wxBitmap (in all ports except for wxMSW) to
implement wxBitmapBase pure virtual functions, but it's implemented only
by wxWidgets itself and shouldn't be overridden again, so declare it as
final in all the ports overriding it and make it non-virtual in wxMSW.

Also, none of the other Create() functions in wxWindow-derived classes
is virtual, so this improves consistency as well.

Finally, this avoids clang-analyzer-optin.cplusplus.VirtualCall warning
from clang-tidy due to calling a virtual function from wxBitmap ctor,
which is the right thing to do only because we really want to call
wxBitmap's own version and not anything defined in a derived class.

Closes #24759.