Inline wxWebRequestCURL::DestroyHeaderList()
This trivial function was called only once and it doesn't seem to be
worth having it, especially because performing cleanup in the dtor makes
the code easier to follow.

This also allows to remove the line resetting m_headerList to null
pointer, as it's now clear that it's not needed.

No real changes.