Replace wxWebResponseImpl::Init() with InitFileStorage()
The old function name wasn't very descriptive and even misleading, as
Init() is typically used for simple initializers in wxWidgets, but this
function was anything but, because it could call SetState() in case of
error.

Remove this side effect and just return a Result from it and handle it
in the caller.

This also probably fixes a bug in NSURLSession backend, as it simply
ignored the failure in Init() before and continued with the request
normally instead of abandoning it.