Forbid using wxFileDialogs on Windows from secure desktop
When running on the secure desktop (which doesn't happen for normal
desktop applications but can happen for DLLs loaded by the login
screen), accessing the file system is a security disk but could happen
due to showing the log dialog which provides a "Save" button opening the
file dialog.

Prevent this from happening by adding a check to wxFileDialog verifying
that it doesn't run on the secure desktop.

Co-authored-by: Vadim Zeitlin <vadim@wxwidgets.org>