Add wxWebSessionSync and wxWebRequestSync stabs
This commit just defines the API and adds the documentation.

Note that we avoid documenting wxWebRequestBase and wxWebRequestBase
classes that are implementation details, even if it means that we
duplicate the descriptions of most of the functions between the async
and sync versions of the classes -- but this is actually useful as it
allows to document things working differently in the sync case.