Stop using wxWebAuthChallenge with wxWebRequestSync
The wxWebAuthChallenge-based approach seems to be unnecessarily
complicated for the synchronous use case and isn't supported by the
current NSURLSession backend implementation (although this could be
changed by rewriting it to use a different delegate for the sync
requests and remembering wxWebAuthChallenge in it between the calls to
Execute()).

For now, just drop support for it and require specifying the
authentication information in the URL itself.