Fix library naming when cross compiling for Windows using CMake
When using a cross compilation toolchain, the CMake variable `WIN32`
will evaluate to true and so will `CMAKE_CROSSCOMPILING`. This means
wxWidgets is built using the `WIN32_MSVC_NAMING`, even though it is
built on Linux.

In addition, the FindwxWidgets.cmake file, provided in the CMake
installation checks for `CMAKE_CROSSCOMPILING` to set the
`wxWidgets_FIND_STYLE`. We therefore end up with CMake trying to use the
Unix find-style (through wx-config) to locate a Windows-style wxWidgets
build, which fails.

Fix the issue by simply adding the check for `NOT CMAKE_CROSSCOMPILING`
when determining the naming in the functions.cmake file.

Closes #24761.