Use the same logic for HTTP methods in all backends
No real changes, just add GetHTTPMethod() helper function in the base
wxWebRequestImpl class and use it in all backends: previously, WinHTTP
and NSURLSession ones already did the same thing, but each in their own
way, while libcurl backend used a different logic which was mostly
equivalent, but this was not really obvious, while now it should be.

Also add a test for using a custom method (DELETE), for now without
body, as this is (still) not supported by libcurl backend.