Fix background of transparent controls in wxAuiToolBar in wxMSW
Just painting the background doesn't make it "show through" the controls
with transparent area, such as wxStaticText or wxCheckBox, in wxMSW and
the special wxCustomBackgroundWindow base class must be used for this,
so derive wxAuiToolBar from it to fix background of custom controls in
it.

Add an example of a wxStaticText in a wxAuiToolBar to show that this
works correctly now.

Closes #24765.