Compute wxStaticText best size ourselves if GTK does it wrongly
GTK doesn't take into account changes to the font of a hidden
wxStaticText and returns wrong preferred (i.e. best) size for it which
still uses the old font in this case, which results in all sorts of
problems when this wrong best size is used for the layout.

Work around it by computing the best size ourselves in this case using
the correct font. This seems to return exactly the same results as GTK
returns as long as the font is not changed, but also works correctly
when it is.

This commit is best viewed ignoring whitespace-only changes.

See #16088.