Allow opting out of catching unhandled C++ exceptions
Add a system option to control whether the main loop and threads are run
from within a try-catch block, or outside of it.

Traditionally, wxWidgets has caught all unhandled C++ exceptions
that are about to cause the program to be aborted.

For many situations this is probably a good thing, but the downside
is that it severely complicates debugging the source of such exceptions.
The backtrace will only show the stack since wx's exception handler,
but not where the exception occurred in the first place.

Any crashdumps created by the OS also suffer from the same issue;
backtraces are mostly useless in these cases.

The following classes with try-catch blocks do not obey the new system
option for now: wxDataViewRendererBase, wxDocTemplate, wxIDropTarget