Avoid trailing space with wx-config --c[xx]flags
When WXCONFIG_C[XX]FLAGS is empty, this ends up being:
`echo $._cppflags ""`
which results in a trailing space being printed. This is a minor issue
but it may confuse a tool trying to parse the output of wx-config.

Removing the quotes would instead give `echo $._cppflags `, which does
not print a trailing space.

Closes #24812.