Add RAII helper for saving/restoring Cairo context
This is safer than calling cairo_save() and cairo_restore() manually and
can also be extended to do more things, see the upcoming commit.

There are almost no changes in this commit, but CalcBoundingBox() is now
called before restoring the context in DoDrawEllipticArc() rather than
after. It is not clear if putting this call after cairo_restore() in the
old code was intentional, but it seems wrong and is definitely
inconsistent with all the other similar calls, so this is probably a bug
fix.