Fix using wrong colour in wxGTK wxPrinterDC
The logic for optimizing the colour changes was wrong as the stored
colour values could get out of sync with the values actually used by
Cairo after cairo_restore() was called to undo the changes to the
latter, but not to the former.

Restore m_currentSourceColour when restoring the context too, to prevent
this from happening.

Closes #24729.