wxQt: Safely remove the model associated with the QTreeView, 2nd attempt
Although this should have been fixed by commit 8cb9adb (Safely remove the model associated
with QTreeView in wxQt), it still crashes on Windows in the HandleDestroyedSignal() handler.

Really fix this now by calling ClearAll() before removing the model, and explicitly delete
the selection model which is not owned by the QTreeView control, according to the Qt docs.