Enable the possibility to warn about unused wxStrings
Use the wxWARN_UNUSED attribute in wxString's definition. Compilers do not
typically warn about unused non-POD objects, but this attribute makes it
possible (on GCC and Clang).

Whether a warning is actually shown or not still depends on whether
the user has enabled warnings for unused variables in their compiler flags.