Add unit test checking the wxNumValidator normalization fix
Verify that a control with wxNumValidator using wxNUM_VAL_ZERO_AS_BLANK
style doesn't change when it loses focus.

This test would have failed without the fix in the parent commit,
although we have to mark the control as dirty to see it to avoid this
from being hidden by the changes done in 67f634dfc4 (Don't normalize the
value of empty text control on focus loss, 2023-07-02).