Fix regression when using initially 0-sized wxAuiNotebook pages
After the changes in 32f3dbd8f4 (Fix background of transparent controls
in wxAuiToolBar in wxMSW, 2024-08-23) creating a 0-sized page resulted
in asserts and crashes due to failure to create 0-sized bitmap.

Don't even try to create such bitmap, as this is indeed invalid, and
just wait until the window has a valid size.

Closes #24808.