Avoid infinite recursion in wxDefaultAssertHandler() more reliably
The changes of 4d1ea475c8 (Add a system option to exit immediately on
assert failure., 2010-06-20) added a wxSystemOptions::GetOptionInt()
call before checking for reentry into the assert handler, which could
(and did, see the scenario described in the parent commit message)
result in an infinite recursion.

Ensure that this doesn't happen by checking for the recursion first, and
only then checking for the system option: if it is set, it doesn't
change anything because we can't ever reenter the assert handler and if
it's not set, it changes even less as we don't do anything in this case.