wxIPCMessageBase serves as the new socket interface.
- The IPCCode is expanded from 8 bits to 32 bits. The first 24 bits
serve as a constant header to ensure that data sync is not lost.
Before this change, loss of sync meant there was an 11 out of 256
chance that an incoming byte would be interpreted as a command,
possibly leading to a crash. The header reduces the probability by a
factor of 2^24.

- Each derived message (Request, Execute, etc) will have a
DataToSocket() and DataFromSocket() method. These methods allow for
easy verification that the reads and writes are correctly symmetric.

- Pointers to Read and write data are separate member vars, for easier
management of const-params for the wxConnection interface (which we
take to be immutable).

- A pointer back to wxTCPEventHandler is going to be needed for
memory management. More on this in a later commit.