Create an array to track memory allocation.
One thing in the previous version: Multiple threads requesting
information would run into problems if data reads overlapped, perhaps
in multiple threads.

The source of this problem is that the memory management from data
reads in wxConnection is not transferred to to the caller. So
wxConnection uses a buffer that gets overwritten for every read.

My approach is to create an array of char* pointers: Every memory
allocation is stored in the array, and survives until another
MAX_MSG_BUFFERS is allocated, at which point the previous data is
deleted and a new allocation occurs.

The deletion of previous data is still potentially problematic (the
caller needs to understand that passed data must be copied in a
reasonable amount of time if they plan to keep it around), but at
least the data will have a lifetime greater than one read.

I set the array to 2048 blocks, which was more for my app than
anything, but maybe there's a better value.