Stop throwing away socket read notifications.
FD_CONNECT, FD_WRITE, FD_READ are non-exclusive notifications in
theory, ie there is a notification with an or'd FD_CONNECT|FD_READ. By
throwing this away, there are occasional stoppages in the data flow,
particularly right after connect.

In any case, I don't understand why sockmsw was throwing any
notification away. It should be up to the callback whether or not
action is to be taken on a notification.  As such, I've eliminated
the filter.