Fix always using "All pages" in wxGTK wxPrintDialog
The code in wxGtkPrinter tried to determine whether all pages should be
printed from the values returned by GetPageInfo(), but this doesn't make
much sense because a typical implementation of GetPageInfo() uses the
values in the associated wxPrintDialogData to determine the pages to
return, i.e. it's wxPrintDialogData which is the primary source of truth
and the values set in it should be respected, otherwise we ended up with
"All pages" being selected in the print dialog independently of the
values set in the application code.

Just remove this code, which isn't present in the other ports neither,
to fix this.

Closes #24868.