Update docks sizes before serializing them
This ensures that restoring the layout by recreating the same docks
recreates the same geometry as was used before, while without it the
panes just got their default initial sizes instead.

Ideal would be to avoid serializing the docks at all, as they can be
recreated automatically to accommodate the panes, but currently there is
no way to specify the size for the docks being recreated, so instead we
have to recreate it with the correct size ourselves when deserializing.