Make FindDocks() helper more ergonomic and a bit more efficient
Return the array from it instead of taking it as output parameter and
add a new "flags" parameter which can be used to tell the function that
it's enough to find just the first dock, instead of getting all of them,
as this is unnecessary if the caller only uses the first element of the
returned array anyhow.

Returning the array makes it possible to use range-for loop to iterate
directly over the result of calling this function, which simplifies the
code in the caller.

To allow doing it even in the places where the code iterated over the
array elements in reverse order, also add FindDocksFlags::ReverseOrder
which tells the function itself to reverse the array. This wouldn't be
necessary if we could use C++20 reverse range adaptor but we're not
quite there yet.

No real changes.