wxQt: Fix wxEVT_{ENTER,LEAVE}_WINDOW events generation
Quoting Qt docs:

- enterEvent() is called when the mouse enters the widget's screen space.
	(This excludes screen space owned by any of the widget's children.)
- leaveEvent() is called when the mouse leaves the widget's screen space.
	If the mouse enters a child widget it will not cause a leaveEvent().

Which contradicts what wx docs says:

For the wxEVT_ENTER_WINDOW and wxEVT_LEAVE_WINDOW events purposes, the mouse
is considered to be inside the window if it is in the window client area and
not inside one of its children. In other words, the parent window receives
wxEVT_LEAVE_WINDOW event not only when the mouse leaves the window entirely but
also when it enters one of its children.