Fix height of wxOwnerDrawnComboBox popup at different DPI
Don't use a hard-coded height of 400px but estimate the height of 21 rows
containing text (the original 400px at default DPI resulted in 21 items).
The actual height of a row depends on it's content and (custom) OnGetRowHeight
implementation and is not known, and can also be different per row.

Also make the height DPI aware when no preferred height is given,
and when the popup is empty.

Fixes #24850