Change wxSizer::Detach(wxWindow*) to take wxWindowBase* instead
This function is called from wxWindowBase dtor to detach the window
being destroyed from the containing sizer and casting the pointer to the
object to wxWindow is UB because the object is not actually a wxWindow
any more by then.

Remove the cast and pass wxWindowBase pointer to wxSizer::Detach()
instead, as it doesn't need the full wxWindow anyhow.

Also document that any wxSizer-derived classes overriding Detach() will
need to be modified to follow this change.