Avoid undefined behaviour in wxScrollHelperBase ctor
Don't cast the not yet fully constructed object to wxScrollHelper when
passing it to wxWindow::SetScrollHelper() and call the latter in the
derived class ctors instead.

An alternative solution could be to store wxScrollHelperBase pointer in
wxWindow instead of wxScrollHelper, as this would be enough, but this
seems a bit ugly, as "Base" classes are supposed to be implementation
detail more than anything else, and duplicating SetScrollHelper() call
in both generic and GTK implementation doesn't seem to be that bad in
comparison.