Fix incorrect cast in wxPGComboBoxEditor::OnEvent()
Unlike the fixes in the precious commits, this was not "just" UB but
could easily result in a crash because the object being cast to
wxOwnerDrawnComboBox was not actually of this type at all (but a
wxComboCtrlTextCtrl), so check whether we have a wxOwnerDrawnComboBox
before using it.

There should probably be a better fix for this, e.g. by checking the
event and then only doing this if the type of the object associated with
the event is really wxOwnerDrawnComboBox, but for now this fixes a bad
problem and so is better than nothing.