* set default infobar BG and FG colors only if not set
We cannot set the background and foreground colors of the infobar before we create the infobar because these colors are set in the method for creating the infobar. If we change them later, then the close button will not fit well because it uses the background color of the infobar we want to change, so we would also need to change the background color of the close button, but there is no function to get the button. Simple solution is to just set FG and BG color if this hasn't happened yet.